Backed out changeset 1091690427f2 (bug 1555644) for causing browser-chrome failures. CLOSED TREE
UltraBlame original commit: 4411daffea4e597e7c845944db67d8da3b377556

+++ toolkit/components/osfile/modules/osfile_unix_front.jsm
@@ -722,6 +722,12 @@
       
       
       
+      
+      if (ctypes.errno != Const.EXDEV || options.noCopy) {
+        throw new File.Error("move", ctypes.errno, sourcePath);
+      }
+
+      
       File.copy(sourcePath, destPath, options);
       
       File.remove(sourcePath);
